US2243: Added more tests for SHNTemperatureMeasurement
git-svn-id: https://redmine.natlab.research.philips.com/svnroot/moonshine/trunk/03_software/Source/Android/Moonshine/shinelib@4444 3c2db180-f856-4c6d-be1c-585a5d938589

+++ src/main/java/com/pins/philips/shinelib/services/healththermometer/SHNTemperatureMeasurement.java
@@ -60,7 +60,7 @@ public class SHNTemperatureMeasurement {
         float result = Float.NaN;
         int value = byteBuffer.getInt();
         if (value != 0x007FFFFF) { // Literal value for NaN in IEEE 11073
-            int exponent = (value >> 24) & 0xFF;
+            int exponent = (int)(byte)((value >> 24) & 0xFF);
             int mantissa = (value & 0x00FFFFFF);
             result = (float) (mantissa * Math.pow(10, exponent));
         }
